+++ R/geom_feature.R
@@ -26,6 +26,14 @@
 #' - linewidth (the former size aesthetic has been deprecated and will be
 #' removed in future versions)
 #'
+#' Prior to version 0.6.0.9001, linewidth was expressed in points, not millimetres,
+#' with a default value of 1. This was inconsistent with both
+#' `geom_gene_arrow()` and ggplot2 convention. From version 0.6.0.9001, linewidth
+#' is expressed in millimetres, and the default value is 0.3. This results in visually
+#' near-identical linewidths if using the default, but may result in a
+#' significant change in linewidths if this value is set. To correct for this
+#' change, divide previous linewidth values by `ggplot2::.pt`.
+#'
 #' @param mapping,data,stat,position,na.rm,show.legend,inherit.aes,... As
 #' standard for ggplot2. inherit.aes is set to FALSE by default, as features
 #' are not likely to share any plot aesthetics other than y.
@@ -44,7 +52,7 @@
 #' ggplot2::ggplot(example_genes, ggplot2::aes(xmin = start, xmax = end,
 #'                                             y = molecule, fill = gene)) +
 #'   geom_gene_arrow() +
-#'   geom_feature(data = example_features, ggplot2::aes(x = position, y = molecule, 
+#'   geom_feature(data = example_features, ggplot2::aes(x = position, y = molecule,
 #'                                                      forward = forward)) +
 #'   ggplot2::facet_wrap(~ molecule, scales = "free")
 #'
@@ -84,23 +92,24 @@ geom_feature <- function(
 
 #' GeomFeature
 #' @noRd
-GeomFeature <- ggplot2::ggproto("GeomFeature", ggplot2::Geom,
+GeomFeature <- ggplot2::ggproto(
+  "GeomFeature",
+  ggplot2::Geom,
   required_aes = c("x", "y"),
   default_aes = ggplot2::aes(
     forward = NA,
     alpha = 1,
     colour = "black",
     linetype = 1,
-    linewidth = 1
+    linewidth = 0.3
   ),
 
   draw_key = ggplot2::draw_key_abline,
 
   setup_data = function(data, params) {
-
     # The 'forward' aesthetic, if provided, should be logical or coerced to
     # logical
-    if (! is.null(data$forward)) {
+    if (!is.null(data$forward)) {
       data$forward <- as.logical(data$forward)
     }
 
@@ -115,7 +124,6 @@ GeomFeature <- ggplot2::ggproto("GeomFeature", ggplot2::Geom,
     feature_width,
     arrowhead_width
   ) {
-
     # Detect coordinate system and transform values
     coord_system <- get_coord_system(coord)
     data <- data_to_grid(data, coord_system, panel_scales, coord)
@@ -138,39 +146,57 @@ GeomFeature <- ggplot2::ggproto("GeomFeature", ggplot2::Geom,
 #' @importFrom grid makeContent
 #' @export
 makeContent.featuretree <- function(x) {
-
   data <- x$data
 
   # Prepare grob for each feature
   grobs <- lapply(seq_len(nrow(data)), function(i) {
-
     feature <- data[i, ]
 
     # Set up geometry
     r <- ifelse(x$coord_system == "polar", feature$away, NA)
-    feature_alongness <- unit_to_alaw(x$feature_width, "along", x$coord_system, r)
-    arrowhead_alongness <- unit_to_alaw(x$arrowhead_width, "along", x$coord_system, r)
-    feature_awayness <- unit_to_alaw(x$feature_height, "away", x$coord_system, r)
+    feature_alongness <- unit_to_alaw(
+      x$feature_width,
+      "along",
+      x$coord_system,
+      r
+    )
+    arrowhead_alongness <- unit_to_alaw(
+      x$arrowhead_width,
+      "along",
+      x$coord_system,
+      r
+    )
+    feature_awayness <- unit_to_alaw(
+      x$feature_height,
+      "away",
+      x$coord_system,
+      r
+    )
 
     # Determine whether this is a feature with orientation or not (i.e. whether
     # or not to draw an elbow and arrowhead), and generate appropriate polyline
 
     # For non-oriented features:
-    if (is.na(feature$forward) | ! is.logical(feature$forward)) {
-
+    if (is.na(feature$forward) | !is.logical(feature$forward)) {
       alongs <- c(feature$along, feature$along)
       aways <- c(feature$away, feature$away + feature_awayness)
       arrow <- NULL
 
-    # For oriented features:
+      # For oriented features:
     } else {
-
       arrow_sign <- ifelse(feature$forward, 1, -1)
       end_along <- feature$along + (feature_alongness * arrow_sign)
       alongs <- c(feature$along, feature$along, end_along)
-      aways <- c(feature$away, feature$away + feature_awayness, 
-                 feature$away + feature_awayness)
-      arrow <- grid::arrow(angle = 20, length = x$arrowhead_width, type = "closed")
+      aways <- c(
+        feature$away,
+        feature$away + feature_awayness,
+        feature$away + feature_awayness
+      )
+      arrow <- grid::arrow(
+        angle = 20,
+        length = x$arrowhead_width,
+        type = "closed"
+      )
     }
 
     # If in polar coordinates, segment the polyline
@@ -182,7 +208,7 @@ makeContent.featuretree <- function(x) {
 
     # Convert polyline into Cartesian coordinates within the grid viewport
     coords <- alaw_to_grid(alongs, aways, x$coord_system, r)
-    
+
     # Generate polyline grob for the feature
     pg <- grid::polylineGrob(
       x = coords$x,
@@ -192,7 +218,9 @@ makeContent.featuretree <- function(x) {
         col = feature$colour,
         fill = feature$colour,
         lty = feature$linetype,
-        lwd = (feature$linewidth %||% feature$size)
+        # linewidth is expressed in mm but grid expects points; multiplying by
+        # .pt converts
+        lwd = (feature$linewidth %||% feature$size) * ggplot2::.pt
       )
     )